requirements.txt
accelerate==0.21.0
bitsandbytes==0.41.1
datasets==2.12.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
evaluate==0.4.1
fastapi==0.109.1
huggingface-hub==0.22.2
Jinja2==3.1.3
numpy==1.24.3
openai==0.27.7
peft @ git+https://github.com/huggingface/peft.git@412d7bc985f52225be8d9a7abc6a0cf2d71b2415
PyYAML==6.0
requests==2.31.0
responses==0.18.0
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
tiktoken==0.6.0
tokenizers==0.15.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cu121
transformers==4.39.3
